Fix edge case bug in `extractIds`
