+++ backend/src/middleware/notifications/extractIds/index.js
@@ -2,6 +2,7 @@ import cheerio from 'cheerio'
 const ID_REGEX = /\/profile\/([\w\-.!~*'"(),]+)/g
 
 export default function (content) {
+  if (!content) return []
   const $ = cheerio.load(content)
   const urls = $('.mention').map((_, el) => {
     return $(el).attr('href')
